Melengkapi Approval

+++ application/controllers/Request_form.php
@@ -32,14 +32,18 @@ class Request_form extends CI_Controller
         $row = $this->Request_form_model->get_by_id(decrypt_url($id));
         if ($row) {
             $data = array(
-		'request_form_id' => $row->request_form_id,
-        'sett_apps' =>$this->Setting_app_model->get_by_id(1),
-		'kode_request_form' => $row->kode_request_form,
-		'user_id' => $row->user_id,
-		'tanggal_request' => $row->tanggal_request,
-		'categori_request_id' => $row->categori_request_id,
-		'keterangan' => $row->keterangan,
-	    );
+        		'request_form_id' => $row->request_form_id,
+                'sett_apps' =>$this->Setting_app_model->get_by_id(1),
+        		'kode_request_form' => $row->kode_request_form,
+        		'user_id' => $row->user_id,
+        		'tanggal_request' => $row->tanggal_request,
+        		'categori_request_id' => $row->categori_request_id,
+        		'keterangan' => $row->keterangan,
+                'status' => $row->status,
+                'whoisreviewing' => $row->approval,
+                'keterangan_tolak' => $row->keterangan_tolak,
+                'classnyak' => $this
+            );
             $this->template->load('template','request_form/request_form_read', $data);
         } else {
             $this->session->set_flashdata('error', 'Record Not Found');
@@ -52,20 +56,62 @@ class Request_form extends CI_Controller
         is_allowed($this->uri->segment(1),'create');
         $data = array(
             'button' => 'Create',
-            'categori_request_id' =>$this->Categori_request_model->get_all(),
+            'categori_requests' =>$this->Categori_request_model->get_all(),
             'kode'=>$this->Request_form_model->get_no_rf(),
             'sett_apps' =>$this->Setting_app_model->get_by_id(1),
             'action' => site_url('request_form/create_action'),
-	    'request_form_id' => set_value('request_form_id'),
-	    'kode_request_form' => set_value('kode_request_form'),
-	    'user_id' => set_value('user_id'),
-	    'tanggal_request' => set_value('tanggal_request'),
-	    'categori_request_id' => set_value('categori_request_id'),
-	    'keterangan' => set_value('keterangan'),
-	);
+    	    'request_form_id' => set_value('request_form_id'),
+    	    'kode_request_form' => set_value('kode_request_form'),
+    	    'user_id' => set_value('user_id'),
+    	    'tanggal_request' => set_value('tanggal_request'),
+    	    'categori_request_id' => set_value('categori_request_id'),
+    	    'keterangan' => set_value('keterangan'),
+            'classnyak' => $this
+    	);
         $this->template->load('template','request_form/request_form_form', $data);
     }
+
+    function generate_approval_list($id_categori_request)
+    {
+        $data = $this->Categori_request_model->get_all_approval_name_and_step($id_categori_request);
+
+        $arr = [];
+
+        $i = 0;
+        
+        foreach ($data as $key => $value) {
+
+            $arr[$i]['user_id'] = $value->user_id;
+            $arr[$i]['status'] = '-';
+            
+            if ($i == 0) {
+            
+                $arr[$i]['tanda_tangan'] = 'sekarang';
+            } else {
+                $arr[$i]['tanda_tangan'] = 'belum';
+            }
+            
+            $i++;
+        }
+
+        return json_encode($arr);
+
+        //return json_encode($arr);
+
+    }
     
+    function cekDataInApprovalListAvailability()
+    {
+        $data = $this->Categori_request_model->get_request_approve_availability($id);
+        return $data;
+    }
+
+    function find_berkas_for_this_request_form($id,$user_id)
+    {
+        $data = $this->Categori_request_model->get_all_file_for_request_form($id,$user_id);
+        return $data;
+    }
+
     public function create_action() 
     {
         is_allowed($this->uri->segment(1),'create');
@@ -74,38 +120,142 @@ class Request_form extends CI_Controller
         if ($this->form_validation->run() == FALSE) {
             $this->create();
         } else {
+
+            $approval_list = $this->generate_approval_list($this->input->post('categori_request_id',TRUE));
+
+            // foreach($approval_list as $k)
+            // {
+            //     echo $k['tanda_tangan'];
+            // }
+
             $data = array(
-		'kode_request_form' => $this->input->post('kode_request_form',TRUE),
-		'user_id' => $this->input->post('user_id',TRUE),
-		'tanggal_request' => $this->input->post('tanggal_request',TRUE),
-		'categori_request_id' => $this->input->post('categori_request_id',TRUE),
-		'keterangan' => $this->input->post('keterangan',TRUE),
-	    );
+        		'kode_request_form' => $this->input->post('kode_request_form',TRUE),
+        		'user_id' => $this->input->post('user_id',TRUE),
+        		'tanggal_request' => $this->input->post('tanggal_request',TRUE),
+        		'categori_request_id' => $this->input->post('categori_request_id',TRUE),
+        		'keterangan' => $this->input->post('keterangan',TRUE),
+                'status' => 'Dalam Review',
+                'approval' => $approval_list,
+                'keterangan_tolak' => '-'
+    	    );
 
             $this->Request_form_model->insert($data);
-            $this->session->set_flashdata('message', 'Create Record Success');
-            redirect(site_url('request_form'));
+            
+            $nama_berkas = $_POST['nama_berkas'];
+
+            if ($nama_berkas) {
+                $this->load->library('upload');
+                
+                $jumlah_data = count($nama_berkas);
+
+                for($i = 0; $i < $jumlah_data;$i++)
+                {
+                    if($_FILES['berkas']['name'][$i]){
+                        
+                        $filenamee = 'ApprDoc-'.$this->input->post('kode_request_form').'-'.date('ymd').'-'.substr(sha1(rand()),0,10);
+
+                        $config['upload_path']          = './assets/assets/img/berkas'; 
+                        $config['allowed_types']        = 'jpg|png|pdf|docx|doc';
+                        $config['max_size']             = 10000;
+                        $config['file_name']            = $filenamee;
+
+                        $_FILES['file[]']['name'] = $_FILES['berkas']['name'][$i];
+                        $_FILES['file[]']['type'] = $_FILES['berkas']['type'][$i];
+                        $_FILES['file[]']['tmp_name'] = $_FILES['berkas']['tmp_name'][$i];
+                        $_FILES['file[]']['error'] = $_FILES['berkas']['error'][$i];
+                        $_FILES['file[]']['size'] = $_FILES['berkas']['size'][$i];
+
+                        $this->upload->initialize($config);
+                        $this->upload->do_upload('file[]');
+
+                        $uploadData[] = $this->upload->data();
+
+                        $data = array(
+                            'karyawan_id' => $this->session->userdata('userid'),
+                            'nama_berkas' => $nama_berkas[$i],
+                            'photo' => $uploadData[$i]['file_name'],
+                        );
+
+                        print_r($data);
+                        
+                        $this->db->insert('berkas',$data);
+                    
+                    } else {
+
+                        echo "haha! no files for ".$nama_berkas[$i].'???'.$_FILES['berkas']['name'][$i].'???';
+                    }
+
+                }
+           }
+
+
+           $this->session->set_flashdata('message', 'Create Record Success');
+           redirect(site_url('request_form'));
         }
     }
+
+    function getusername($id)
+    {
+        $this->load->model('User_m');
+        $data = $this->User_m->get_by_id($id);
+        return $data;
+    }
     
+    function is_allowed_toedit($id_request_form)
+    {
+        $data = $this->Request_form_model->get_by_id(decrypt_url($id_request_form));
+
+        if($data->status == 'Dalam Review')
+        {
+            $arr = json_decode($data->approval, TRUE);
+
+            $stillonreview = 0;
+
+            foreach ($arr as $value) {
+                if ($value['status'] === '-') {
+                    $stillonreview++;
+                }
+            }
+
+            if ($stillonreview == count($arr)) {
+                return 'allowed';
+            } else {
+                return 'not allowed';
+            }
+        }
+
+        if ($data->status == 'Ditolak') {
+            return 'allowed';
+        }
+    }
+
     public function update($id) 
     {
         is_allowed($this->uri->segment(1),'update');
+
         $row = $this->Request_form_model->get_by_id(decrypt_url($id));
 
         if ($row) {
-            $data = array(
-                'button' => 'Update',
-                'sett_apps' =>$this->Setting_app_model->get_by_id(1),
-                'action' => site_url('request_form/update_action'),
-		'request_form_id' => set_value('request_form_id', $row->request_form_id),
-		'kode_request_form' => set_value('kode_request_form', $row->kode_request_form),
-		'user_id' => set_value('user_id', $row->user_id),
-		'tanggal_request' => set_value('tanggal_request', $row->tanggal_request),
-		'categori_request_id' => set_value('categori_request_id', $row->categori_request_id),
-		'keterangan' => set_value('keterangan', $row->keterangan),
-	    );
-            $this->template->load('template','request_form/request_form_form', $data);
+
+            if ($this->is_allowed_toedit($id) == 'allowed') {
+                $data = array(
+                    'button' => 'Update',
+                    'sett_apps' =>$this->Setting_app_model->get_by_id(1),
+                    'action' => site_url('request_form/update_action'),
+                    'request_form_id' => set_value('request_form_id', $row->request_form_id),
+                    'kode_request_form' => set_value('kode_request_form', $row->kode_request_form),
+                    'user_id' => set_value('user_id', $row->user_id),
+                    'tanggal_request' => set_value('tanggal_request', $row->tanggal_request),
+                    'categori_request_id' => $row->categori_request_id,
+                    'keterangan' => set_value('keterangan', $row->keterangan),
+                    'keterangan_tolak' => $row->keterangan_tolak,
+                    'classnyak' => $this
+                );
+                $this->template->load('template','request_form/request_form_form', $data);
+            } else {
+                $this->session->set_flashdata('message', 'Tidak dapat diedit karena sudah direview');
+                redirect(site_url('request_form'));
+            }
         } else {
             $this->session->set_flashdata('error', 'Record Not Found');
             redirect(site_url('request_form'));
@@ -120,15 +270,72 @@ class Request_form extends CI_Controller
         if ($this->form_validation->run() == FALSE) {
             $this->update($this->input->post('request_form_id', TRUE));
         } else {
-            $data = array(
-		'kode_request_form' => $this->input->post('kode_request_form',TRUE),
-		'user_id' => $this->input->post('user_id',TRUE),
-		'tanggal_request' => $this->input->post('tanggal_request',TRUE),
-		'categori_request_id' => $this->input->post('categori_request_id',TRUE),
-		'keterangan' => $this->input->post('keterangan',TRUE),
-	    );
+            $row = $this->Request_form_model->get_by_id($this->input->post('request_form_id', TRUE));
 
-            $this->Request_form_model->update($this->input->post('request_form_id', TRUE), $data);
+            if ($row->status == 'Ditolak' || $row->status == 'Dalam Review') {
+                $data = array(
+                    'kode_request_form' => $this->input->post('kode_request_form',TRUE),
+                    'user_id' => $this->input->post('user_id',TRUE),
+                    'tanggal_request' => $this->input->post('tanggal_request',TRUE),
+                    'categori_request_id' => $this->input->post('categori_request_id',TRUE),
+                    'keterangan' => $this->input->post('keterangan',TRUE),
+                    'approval' => $this->generate_approval_list($this->input->post('categori_request_id',TRUE)),
+                    'status' => 'Dalam Review'
+                );
+                //print_r($data);
+                $this->Request_form_model->update($this->input->post('request_form_id', TRUE), $data);
+            }
+
+            // IN CASE YOUR ADDING AGAIN
+
+            $nama_berkas = $_POST['nama_berkas'];
+
+            if ($nama_berkas) {
+                $this->load->library('upload');
+                
+                $jumlah_data = count($nama_berkas);
+
+                for($i = 0; $i < $jumlah_data;$i++)
+                {
+                    if($_FILES['berkas']['name'][$i]){
+                        
+                        $filenamee = 'ApprDoc-'.$this->input->post('kode_request_form').'-'.date('ymd').'-'.substr(sha1(rand()),0,10);
+
+                        $config['upload_path']          = './assets/assets/img/berkas'; 
+                        $config['allowed_types']        = 'jpg|png|pdf|docx|doc';
+                        $config['max_size']             = 10000;
+                        $config['file_name']            = $filenamee;
+
+                        $_FILES['file[]']['name'] = $_FILES['berkas']['name'][$i];
+                        $_FILES['file[]']['type'] = $_FILES['berkas']['type'][$i];
+                        $_FILES['file[]']['tmp_name'] = $_FILES['berkas']['tmp_name'][$i];
+                        $_FILES['file[]']['error'] = $_FILES['berkas']['error'][$i];
+                        $_FILES['file[]']['size'] = $_FILES['berkas']['size'][$i];
+
+                        $this->upload->initialize($config);
+                        $this->upload->do_upload('file[]');
+
+                        $uploadData[] = $this->upload->data();
+
+                        $data = array(
+                            'karyawan_id' => $this->session->userdata('userid'),
+                            'nama_berkas' => $nama_berkas[$i],
+                            'photo' => $uploadData[$i]['file_name'],
+                        );
+
+                        print_r($data);
+                        
+                        $this->db->insert('berkas',$data);
+                    
+                    } else {
+
+                        echo "haha! no files for ".$nama_berkas[$i].'???'.$_FILES['berkas']['name'][$i].'???';
+                    }
+
+                }
+            }
+
+
             $this->session->set_flashdata('message', 'Update Record Success');
             redirect(site_url('request_form'));
         }
@@ -140,13 +347,44 @@ class Request_form extends CI_Controller
         $row = $this->Request_form_model->get_by_id(decrypt_url($id));
 
         if ($row) {
-            $this->Request_form_model->delete(decrypt_url($id));
-            $this->session->set_flashdata('message', 'Delete Record Success');
-            redirect(site_url('request_form'));
+
+            if ($this->is_allowed_toedit($id) == 'allowed') {
+                $getberkas = $this->Request_form_model->get_berkas_list($row->kode_request_form,$row->user_id);
+
+                foreach ($getberkas as $value) {
+                    $this->Request_form_model->delete_berkas_form_request($value->berkas_id);
+                    unlink('./assets/assets/img/berkas/'.$value->photo);
+                }
+
+                $this->Request_form_model->delete(decrypt_url($id));
+
+                $this->session->set_flashdata('message', 'Delete Record Success');
+                redirect(site_url('request_form'));
+            }
+            else
+            {
+                $this->session->set_flashdata('message', 'Tidak dapat dihapus karena sudah dalam proses review');
+                redirect(site_url('request_form'));   
+            }
+            
         } else {
             $this->session->set_flashdata('error', 'Record Not Found');
             redirect(site_url('request_form'));
         }
+    }
+
+    public function delete_berkas()
+    {
+        is_allowed($this->uri->segment(1),'delete');
+
+        $id = $this->input->post('berkas_id');
+        $filename = $this->input->post('file_name');
+
+        $this->Request_form_model->delete_berkas_form_request(decrypt_url($id));
+        
+        unlink("./assets/assets/img/berkas/".$filename);
+
+        echo 'ok';
     }
 
     public function _rules()
